Added test cases for nested function calls.
Currently, nested function calls are not parsed correctly, but this can
be worked around by introducing variable, keeping result of nested call.

It turns out that this workaround is possible because of bug in code making
variable expansion which recognizes variable "$darkBlue," in this test.
Following evaluation of darken() detects two arguments "$darkBlue" and
"1%", where variable expansion works correctly.

We should take care not to break this workaround when fixing the parser.